+++ packages/use/src/useCountdown/index.ts
@@ -1,5 +1,5 @@
-import type { MaybeRefOrGetter, Ref } from 'vue-demi';
-import type { TransitionOptions } from '@vueuse/core';
+import type { Ref } from 'vue-demi';
+import type { MaybeRefOrGetter, TransitionOptions } from '@vueuse/core';
 import { whenever, wheneverEffectScopeImmediate } from '@mixte/use';
 import { TransitionPresets, executeTransition, syncRef, toValue, tryOnScopeDispose } from '@vueuse/core';
 import { computed, ref, watch } from 'vue-demi';